+++ revelare/utils_revelare/clean_data.py
@@ -4,10 +4,6 @@ import frappe
 from frappe import _
 import json
 
-# Permite trabajar con acentos, ñ, simbolos, etc
-import os, sys
-reload(sys)
-sys.setdefaultencoding('utf-8')
 
 def preparar_data_tabla():
     pass
