+++ revelare/utils_revelare/sales_invoice.py
@@ -1,0 +1,5 @@
+# -*- coding: utf-8 -*-
+from __future__ import unicode_literals
+import frappe
+from frappe import _
+import json
